python/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
pdfplumber>=0.10.0
PyMuPDF>=1.23.0
python-docx>=1.1.0
openai>=1.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
